Add CSRF protection to delete form.

+++ application/Omeka/src/Omeka/Controller/Admin/ItemController.php
@@ -2,6 +2,7 @@
 namespace Omeka\Controller\Admin;
 
 use Omeka\Api\ResponseFilter;
+use Omeka\Form\DeleteForm;
 use Zend\Mvc\Controller\AbstractActionController;
 use Zend\View\Model\ViewModel;
 
@@ -26,8 +27,13 @@ class ItemController extends AbstractActionController
             $this->apiError($response);
             return;
         }
+
         $this->paginator($response->getTotalResults(), $page);
         $view->setVariable('items', $response->getContent());
+        $view->setVariable('deleteForm', new DeleteForm);
+        return $view;
+    }
+
         return $view;
     }
 
@@ -49,14 +55,20 @@ class ItemController extends AbstractActionController
     public function deleteAction()
     {
         if ($this->getRequest()->isPost()) {
-            $response = $this->api()->delete(
-                'items', array('id' => $this->params('id'))
-            );
-            if ($response->isError()) {
-                $this->apiError($response);
-                return;
+            $form = new DeleteForm;
+            $form->setData($this->getRequest()->getPost());
+            if ($form->isValid()) {
+                $response = $this->api()->delete(
+                    'items', array('id' => $this->params('id'))
+                );
+                if ($response->isError()) {
+                    $this->messenger()->addError('Item could not be deleted');
+                } else {
+                    $this->messenger()->addSuccess('Item successfully deleted');
+                }
+            } else {
+                $this->messenger()->addError('Item could not be deleted');
             }
-            $this->messenger()->addSuccess('Item successfully deleted');
         }
         return $this->redirect()->toRoute('admin/default', array(
             'controller' => 'item',
